refactor : move global annotation models to core.shared package
