+++ src/main/java/coz/weavon/core/auth/domain/model/OAuthUser.java
@@ -1,6 +1,6 @@
 package coz.weavon.core.auth.domain.model;
 
-import coz.weavon.common.domain.model.Property;
+import coz.weavon.core.shared.domain.model.Property;
 import coz.weavon.core.auth.domain.service.OAuthUserProvider;
 import java.util.Map;
 import lombok.Builder;
